+++ core/src/main/java/io/substrait/extension/ExtensionCollector.java
@@ -3,7 +3,6 @@ package io.substrait.extension;
 import io.substrait.proto.ExtendedExpression;
 import io.substrait.proto.Plan;
 import io.substrait.proto.SimpleExtensionDeclaration;
-import io.substrait.proto.SimpleExtensionURI;
 import io.substrait.proto.SimpleExtensionURN;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -20,14 +19,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class ExtensionCollector extends AbstractExtensionLookup {
   private final BidiMap<Integer, SimpleExtension.FunctionAnchor> funcMap;
   private final BidiMap<Integer, SimpleExtension.TypeAnchor> typeMap;
-  private final SimpleExtension.ExtensionCollection extensionCollection;
 
   // start at 0 to make sure functionAnchors start with 1 according to spec
   private int counter = 0;
-
-  private String getUriFromUrn(String urn) {
-    return extensionCollection.getUriFromUrn(urn);
-  }
 
   public ExtensionCollector() {
     this(DefaultExtensionCatalog.DEFAULT_COLLECTION);
@@ -40,7 +34,6 @@ public class ExtensionCollector extends AbstractExtensionLookup {
     }
     funcMap = new BidiMap<>(functionAnchorMap);
     typeMap = new BidiMap<>(typeAnchorMap);
-    this.extensionCollection = extensionCollection;
   }
 
   public int getFunctionReference(SimpleExtension.Function declaration) {
@@ -67,7 +60,6 @@ public class ExtensionCollector extends AbstractExtensionLookup {
     SimpleExtensions simpleExtensions = getExtensions();
 
     builder.addAllExtensionUrns(simpleExtensions.urns.values());
-    builder.addAllExtensionUris(simpleExtensions.uris.values());
     builder.addAllExtensions(simpleExtensions.extensionList);
   }
 
@@ -75,22 +67,17 @@ public class ExtensionCollector extends AbstractExtensionLookup {
     SimpleExtensions simpleExtensions = getExtensions();
 
     builder.addAllExtensionUrns(simpleExtensions.urns.values());
-    builder.addAllExtensionUris(simpleExtensions.uris.values());
     builder.addAllExtensions(simpleExtensions.extensionList);
   }
 
   private SimpleExtensions getExtensions() {
     AtomicInteger urnPos = new AtomicInteger(1);
-    AtomicInteger uriPos = new AtomicInteger(1);
     HashMap<String, SimpleExtensionURN> urns = new HashMap<>();
-    HashMap<String, SimpleExtensionURI> uris = new HashMap<>();
 
     ArrayList<SimpleExtensionDeclaration> extensionList = new ArrayList<>();
     for (Map.Entry<Integer, SimpleExtension.FunctionAnchor> e : funcMap.forwardEntrySet()) {
       String urn = e.getValue().urn();
-      String uri = getUriFromUrn(urn);
 
-      // Create URN entry
       SimpleExtensionURN urnObj =
           urns.computeIfAbsent(
               urn,
@@ -100,29 +87,11 @@ public class ExtensionCollector extends AbstractExtensionLookup {
                       .setUrn(k)
                       .build());
 
-      // Create URI entry if mapping exists
-      SimpleExtensionURI uriObj = null;
-      if (uri != null) {
-        uriObj =
-            uris.computeIfAbsent(
-                uri,
-                k ->
-                    SimpleExtensionURI.newBuilder()
-                        .setExtensionUriAnchor(uriPos.getAndIncrement())
-                        .setUri(k)
-                        .build());
-      }
-
-      // Create function declaration with both URN and URI references
       SimpleExtensionDeclaration.ExtensionFunction.Builder funcBuilder =
           SimpleExtensionDeclaration.ExtensionFunction.newBuilder()
               .setFunctionAnchor(e.getKey())
               .setName(e.getValue().key())
               .setExtensionUrnReference(urnObj.getExtensionUrnAnchor());
-
-      if (uriObj != null) {
-        funcBuilder.setExtensionUriReference(uriObj.getExtensionUriAnchor());
-      }
 
       SimpleExtensionDeclaration decl =
           SimpleExtensionDeclaration.newBuilder().setExtensionFunction(funcBuilder).build();
@@ -131,9 +100,7 @@ public class ExtensionCollector extends AbstractExtensionLookup {
 
     for (Map.Entry<Integer, SimpleExtension.TypeAnchor> e : typeMap.forwardEntrySet()) {
       String urn = e.getValue().urn();
-      String uri = getUriFromUrn(urn);
 
-      // Create URN entry
       SimpleExtensionURN urnObj =
           urns.computeIfAbsent(
               urn,
@@ -143,48 +110,27 @@ public class ExtensionCollector extends AbstractExtensionLookup {
                       .setUrn(k)
                       .build());
 
-      // Create URI entry if mapping exists
-      SimpleExtensionURI uriObj = null;
-      if (uri != null) {
-        uriObj =
-            uris.computeIfAbsent(
-                uri,
-                k ->
-                    SimpleExtensionURI.newBuilder()
-                        .setExtensionUriAnchor(uriPos.getAndIncrement())
-                        .setUri(k)
-                        .build());
-      }
-
-      // Create type declaration with both URN and URI references
       SimpleExtensionDeclaration.ExtensionType.Builder typeBuilder =
           SimpleExtensionDeclaration.ExtensionType.newBuilder()
               .setTypeAnchor(e.getKey())
               .setName(e.getValue().key())
               .setExtensionUrnReference(urnObj.getExtensionUrnAnchor());
 
-      if (uriObj != null) {
-        typeBuilder.setExtensionUriReference(uriObj.getExtensionUriAnchor());
-      }
-
       SimpleExtensionDeclaration decl =
           SimpleExtensionDeclaration.newBuilder().setExtensionType(typeBuilder).build();
       extensionList.add(decl);
     }
-    return new SimpleExtensions(urns, uris, extensionList);
+    return new SimpleExtensions(urns, extensionList);
   }
 
   private static final class SimpleExtensions {
     final HashMap<String, SimpleExtensionURN> urns;
-    final HashMap<String, SimpleExtensionURI> uris;
     final ArrayList<SimpleExtensionDeclaration> extensionList;
 
     SimpleExtensions(
         HashMap<String, SimpleExtensionURN> urns,
-        HashMap<String, SimpleExtensionURI> uris,
         ArrayList<SimpleExtensionDeclaration> extensionList) {
       this.urns = urns;
-      this.uris = uris;
       this.extensionList = extensionList;
     }
   }