+++ core/src/main/java/io/substrait/package-info.java
@@ -1,2 +1,6 @@
+/**
+ * The {@code io.substrait} package provides core classes and interfaces for working with the
+ * Substrait specification, which defines a portable representation of query plans and expressions.
+ */
 @org.immutables.value.Value.Style(allowedClasspathAnnotations = {java.lang.Override.class})
 package io.substrait;