+++ isthmus/src/main/java/io/substrait/isthmus/expression/ExpressionRexConverter.java
@@ -54,6 +54,7 @@ import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.sql.SqlOperator;
 import org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.calcite.sql.parser.SqlParserPos;
+import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.util.TimeString;
 import org.apache.calcite.util.TimestampString;
 
@@ -213,9 +214,16 @@ public class ExpressionRexConverter
 
   @Override
   public RexNode visit(PrecisionTimeLiteral expr, Context context) throws RuntimeException {
-    return rexBuilder.makeLiteral(
-        createTimeString(expr.value(), expr.precision()),
-        typeConverter.toCalcite(typeFactory, expr.getType()));
+    int maxPrecision = typeFactory.getTypeSystem().getMaxPrecision(SqlTypeName.TIME);
+    if (expr.precision() > maxPrecision) {
+      throw new IllegalArgumentException(
+          String.format(
+              "unsupported precision_time precision %s, max precision in Calcite type system is set to %s",
+              expr.precision(), maxPrecision));
+    }
+
+    return rexBuilder.makeTimeLiteral(
+        createTimeString(expr.value(), expr.precision()), expr.precision());
   }
 
   protected TimeString createTimeString(long value, int precision) {
@@ -246,7 +254,7 @@ public class ExpressionRexConverter
               .withNanos(fracSecondsInNano);
         }
       default:
-        throw new UnsupportedOperationException(
+        throw new IllegalArgumentException(
             String.format("Cannot handle PrecisionTime with precision %d.", precision));
     }
   }
@@ -279,13 +287,28 @@ public class ExpressionRexConverter
 
   @Override
   public RexNode visit(PrecisionTimestampLiteral expr, Context context) throws RuntimeException {
-    return rexBuilder.makeLiteral(
-        getTimestampString(expr.value(), expr.precision()),
-        typeConverter.toCalcite(typeFactory, expr.getType()));
+    int maxPrecision = typeFactory.getTypeSystem().getMaxPrecision(SqlTypeName.TIMESTAMP);
+    if (expr.precision() > maxPrecision) {
+      throw new IllegalArgumentException(
+          String.format(
+              "unsupported precision_timestamp precision %s, max precision in Calcite type system is set to %s",
+              expr.precision(), maxPrecision));
+    }
+
+    return rexBuilder.makeTimestampLiteral(
+        getTimestampString(expr.value(), expr.precision()), expr.precision());
   }
 
   @Override
   public RexNode visit(PrecisionTimestampTZLiteral expr, Context context) throws RuntimeException {
+    int maxPrecision = typeFactory.getTypeSystem().getMaxPrecision(SqlTypeName.TIMESTAMP_TZ);
+    if (expr.precision() > maxPrecision) {
+      throw new IllegalArgumentException(
+          String.format(
+              "unsupported precision_timestamp precision %s, max precision in Calcite type system is set to %s",
+              expr.precision(), maxPrecision));
+    }
+
     return rexBuilder.makeLiteral(
         getTimestampString(expr.value(), expr.precision()),
         typeConverter.toCalcite(typeFactory, expr.getType()));
@@ -323,7 +346,7 @@ public class ExpressionRexConverter
               .withNanos(fracSecondsInNano);
         }
       default:
-        throw new UnsupportedOperationException(
+        throw new IllegalArgumentException(
             String.format("Cannot handle PrecisionTimestamp with precision %d.", precision));
     }
   }